requirements.txt
openai>=1.0.0
pymupdf>=1.22.0       
python-dotenv>=1.0.0
pytest>=7.0.0
typing-extensions>=4.0.0; python_version < "3.10"
